requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
EasyProcess==1.1
entrypoint2==1.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.52.4
fsspec==2024.6.0
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lapx==0.5.9
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
mpmath==1.3.0
mss==9.0.1
munkres==1.1.4
mutagen==1.47.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
pip==24.0
ply==3.11
protobuf==5.27.1
psutil==5.9.8
py-cpuinfo==9.0.0
pycryptodomex==3.20.0
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyscreenshot==3.1
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.32.2
scipy==1.13.1
seaborn==0.13.2
setuptools==70.0.0
simplejpeg==1.7.2
sip==6.7.12
six==1.16.0
statsmodels==0.14.2
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.24
urllib3==2.2.1
vidgear==0.3.2
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
yt-dlp==2024.5.27
